Moved geolcoation Go SDK to sdk/ rather than smartcontract/sdk
