+++ sdk/geolocation/go/client.go
