+++ sdk/geolocation/go/client_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/gagliardetto/solana-go"
 	solanarpc "github.com/gagliardetto/solana-go/rpc"
-	"github.com/malbeclabs/doublezero/smartcontract/sdk/go/geolocation"
+	geolocation "github.com/malbeclabs/doublezero/sdk/geolocation/go"
 	"github.com/stretchr/testify/require"
 )
