+++ sdk/geolocation/go/constants.go
