+++ sdk/geolocation/go/deserialize.go
