+++ sdk/geolocation/go/main_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/gagliardetto/solana-go"
 	solanarpc "github.com/gagliardetto/solana-go/rpc"
 	"github.com/lmittmann/tint"
-	"github.com/malbeclabs/doublezero/smartcontract/sdk/go/geolocation"
+	geolocation "github.com/malbeclabs/doublezero/sdk/geolocation/go"
 )
 
 var (
