+++ sdk/geolocation/go/state.go
